Reorganise widget files
