+++ app/src/main/kotlin/uk/gov/govuk/widgets/ui/AlertBanner.kt
@@ -1,4 +1,4 @@
-package uk.gov.govuk.ui.widgets
+package uk.gov.govuk.widgets.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
